src/requirements.txt
GitPython
gradio
jupyter
langchain>=0.2.15
langchain-community>=0.2.14
langchain-openai>=0.1.23,
langchain-experimental>=0.0.64
matplotlib
neo4j
numpy
openai>=1.42
pandas
python-dotenv